requirements.txt
# Web & API
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Desktop
pystray>=0.19.0
Pillow>=10.0.0
pywebview>=4.4.0

# Windows
pywin32>=306

# Scheduling & notifications
APScheduler>=3.10.0
desktop-notifier>=3.5.0

# Config & utils
pydantic-settings>=2.0.0
psutil>=5.9.0

# Optional: for running frontend dev (not bundled in exe)
# httpx for tests
httpx>=0.25.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
